Added friendlier error dialogs and confirmation dialogs
...Also,cleaned up the code.

+++ MessageServer/ServerMain.cs
@@ -126,13 +126,10 @@ namespace MessageServer
                 }
                 userList.Close();
             }
-            catch (Exception e)
+            catch
             {
                 Console.WriteLine("Unable to locate file users.dat");
                 Console.WriteLine("Maybe first start or corrupted");
-                Console.WriteLine();
-                Console.WriteLine("Debugging information follows:");
-                Console.WriteLine(e.ToString());
             }
         }
 
@@ -167,19 +164,23 @@ namespace MessageServer
 
         private void buttonClearDatabase_Click(object sender, EventArgs e)  //This clears the database of any records
         {
-            string query = "DELETE FROM [dbo].[Table]";
-            using (SqlConnection connection = new SqlConnection(Properties.Settings.Default.messageDbConnectionString))
+            var confirmationResult = MessageBox.Show("Are you sure you want to clear the database?", "Confirmation", MessageBoxButtons.YesNo);
+            if (confirmationResult == DialogResult.Yes)
             {
-                using (SqlCommand command = new SqlCommand(query, connection))
+                string query = "DELETE FROM [dbo].[Table]";
+                using (SqlConnection connection = new SqlConnection(Properties.Settings.Default.messageDbConnectionString))
                 {
-                    connection.Open();
-                    command.ExecuteNonQuery();
+                    using (SqlCommand command = new SqlCommand(query, connection))
+                    {
+                        connection.Open();
+                        command.ExecuteNonQuery();
+                    }
                 }
-            }
-            this.tableTableAdapter.Fill(this.messageDbDataSet1.Table);
-            foreach (string user in users) //Regenerates the message tables for each user
-            {
-                generateUserMessageFiles(user);
+                this.tableTableAdapter.Fill(this.messageDbDataSet1.Table);
+                foreach (string user in users) //Regenerates the message tables for each user
+                {
+                    generateUserMessageFiles(user);
+                }
             }
         }
     }
